20240920 Update shopping
