+++ src/main/java/com/shopping/bean/ItemBean.java
@@ -1,6 +1,8 @@
 package com.shopping.bean;
 
 
+import org.springframework.stereotype.Component;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -10,6 +12,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
+@Component
 @Entity @Table(name="shopping_item")
 @IdClass(ItemId.class)
 public class ItemBean {
