+++ src/main/java/com/shopping/bean/ProductType.java
@@ -3,6 +3,8 @@ package com.shopping.bean;
 
 import java.util.Set;
 
+import org.springframework.stereotype.Component;
+
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -15,7 +17,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
-
+@Component
 @Entity @Table(name="product_type")
 public class ProductType {
