+++ src/main/java/com/shopping/dao/ItemDao.java
@@ -3,7 +3,11 @@ package com.shopping.dao;
 import java.util.List;
 
 import org.hibernate.Session;
+import org.hibernate.SessionFactory;
 import org.hibernate.query.Query;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.shopping.bean.ItemBean;
 import com.shopping.bean.ItemId;
@@ -11,17 +15,23 @@ import com.shopping.bean.ProductBean;
 import com.shopping.bean.ShoppingBean;
 import com.shopping.bean.ItemBean;
 
+@Repository
+@Transactional
 public class ItemDao {
 
-	private Session session;
+//	private Session session;
 
-	public ItemDao(Session session) {
-		this.session = session;
-	}
+	@Autowired
+	private SessionFactory sessionFactory;
+	
+//	public ItemDao(Session session) {
+//		this.session = session;
+//	}
 
 
 	// 新增訂單明細
 	public ItemBean addItem(ItemBean itembean) {
+		Session session = sessionFactory.getCurrentSession();
 		if (itembean != null) {
 			session.persist(itembean);
 			return itembean;
@@ -31,6 +41,7 @@ public class ItemDao {
 
 	// 刪除單筆
 	public boolean deleteItem(Integer shoppingId, Integer productId) {
+		Session session = sessionFactory.getCurrentSession();
 		ItemId itemId = new ItemId(shoppingId, productId);
 		ItemBean resultBean = session.get(ItemBean.class, itemId);
 		if (resultBean != null) {
@@ -43,6 +54,7 @@ public class ItemDao {
 	
 	// 刪除全部
 	public boolean deleteAllItem(Integer shoppingId) {
+		Session session = sessionFactory.getCurrentSession();
 	    Query<ItemBean> query = session.createQuery("from ItemBean where shoppingId = :shoppingId", ItemBean.class);
 	    query.setParameter("shoppingId", shoppingId);
 	    List<ItemBean> list = query.list();
@@ -60,6 +72,7 @@ public class ItemDao {
 
 	// 更新
 	public ShoppingBean updateProduct(ShoppingBean shoppingBean) {
+		Session session = sessionFactory.getCurrentSession();
 		ShoppingBean isExist = session.get(ShoppingBean.class, shoppingBean.getShoppingId());
 		if (isExist != null) {
 			session.merge(shoppingBean);
@@ -70,10 +83,11 @@ public class ItemDao {
 
 	// 更新產品項目
 	public ItemBean updateItem(Integer shoppingItemQuantity, Integer productId, Integer shoppingId) {
+		Session session = sessionFactory.getCurrentSession();
 		ItemBean item = session
 				.createQuery("FROM ItemBean WHERE shoppingId = :shoppingId AND productId = :productId", ItemBean.class)
 				.setParameter("shoppingId", shoppingId).setParameter("productId", productId).uniqueResult();
-		ProductDao productDao = new ProductDao(session);
+		ProductDao productDao = new ProductDao();
 		Integer productPrice = productDao.searchProductPriceById(productId);
 		Integer itemPrice = shoppingItemQuantity * productPrice;
 		item.setShoppingItemQuantity(shoppingItemQuantity);
@@ -83,6 +97,7 @@ public class ItemDao {
 
 	// 計算訂單總金額
 	public Integer calculateTotalAmount(Integer shoppingId) {
+		Session session = sessionFactory.getCurrentSession();
 		String hql = "SELECT SUM(i.shoppingItemPrice * i.shoppingItemQuantity) FROM ItemBean i WHERE i.shoppingId = :shoppingId";
 		Query<Long> query = session.createQuery(hql, Long.class);
 		query.setParameter("shoppingId", shoppingId);
@@ -92,6 +107,7 @@ public class ItemDao {
 
 	// 查詢訂單明細
 	public List<ItemBean> searchItemsByShoppingId(Integer shoppingId) {
+		Session session = sessionFactory.getCurrentSession();
 		String hql = "FROM ItemBean i WHERE i.shoppingId = :shoppingId";
 		Query<ItemBean> query = session.createQuery(hql, ItemBean.class);
 		query.setParameter("shoppingId", shoppingId);
@@ -100,7 +116,7 @@ public class ItemDao {
 	
 	// 查詢單筆訂單明細
 	public ItemBean searchItem(Integer shoppingId, Integer productId) {
-		
+		Session session = sessionFactory.getCurrentSession();
 		ItemId itemId = new ItemId(shoppingId, productId);
 		ItemBean resultBean = session.get(ItemBean.class, itemId);
 		if (resultBean != null) {
