+++ src/main/java/com/shopping/service/ShoppingService.java
@@ -3,6 +3,9 @@ package com.shopping.service;
 import java.util.List;
 
 import org.hibernate.Session;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.members.bean.Member;
 import com.shopping.bean.ItemBean;
@@ -12,18 +15,21 @@ import com.shopping.dao.ItemDao;
 import com.shopping.dao.ProductDao;
 import com.shopping.dao.ShoppingDao;
 
+@Service
+@Transactional
 public class ShoppingService {
 	
+	@Autowired
 	private ShoppingDao sDao;
 	private ItemDao iDao;
 	private ProductDao pDao;
 	
 	
-	public ShoppingService(Session session) {
-		sDao = new ShoppingDao(session);
-		iDao = new ItemDao(session);
-		pDao = new ProductDao(session);
-	}
+//	public ShoppingService(Session session) {
+//		sDao = new ShoppingDao(session);
+//		iDao = new ItemDao(session);
+//		pDao = new ProductDao(session);
+//	}
 	
 	
 	public ShoppingBean addOrder(ShoppingBean bean) {
