+++ src/main/webapp/Shopping/SearchAllShopping.jsp
@@ -14,7 +14,7 @@
     <script src="../template/template.js"></script>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.datatables.net/1.12.1/js/jquery.dataTables.min.js"></script>
-    <jsp:include page="../HomePage.html" />
+    <jsp:include page="../HomePage.jsp" />
     <meta charset="UTF-8">
     <style>
         .custom-button {
